# backend/swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "API de Usuários",
      "description": "Documentação da API CRUD de usuários desenvolvida com Express e MySQL.",
      "version": "1.0.0"
    },
    "servers": [
      {
        "url": "http://localhost:3000",
        "description": "Servidor Local"
      }
    ],
    "paths": {
      "/usuarios": {
        "post": {
          "summary": "Inserir um novo usuário",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Usuario"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Usuario cadastrado"
            }
          }
        },
        "get": {
          "summary": "Listar todos os usuários",
          "responses": {
            "200": {
              "description": "Lista de usuários",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/UsuarioCompleto"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/usuarios/{id}": {
        "put": {
          "summary": "Atualizar um usuário existente",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer"
              },
              "description": "ID do usuário"
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Usuario"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Usuario atualizado"
            }
          }
        },
        "delete": {
          "summary": "Deletar um usuário",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer"
              },
              "description": "ID do usuário"
            }
          ],
          "responses": {
            "200": {
              "description": "Usuario deletado"
            }
          }
        }
      }
    },
    "components": {
      "schemas": {
        "Usuario": {
          "type": "object",
          "properties": {
            "nome": { "type": "string", "example": "João Silva" },
            "email": { "type": "string", "example": "joao@email.com" },
            "senha": { "type": "string", "example": "123456" }
          }
        },
        "UsuarioCompleto": {
          "type": "object",
          "properties": {
            "id": { "type": "integer", "example": 1 },
            "nome": { "type": "string", "example": "João Silva" },
            "email": { "type": "string", "example": "joao@email.com" },
            "senha": { "type": "string", "example": "123456" }
          }
        }
      }
    }
  }